+++ ubercloneapp/src/main/java/com/github/k2ocabhinav/ubercloneapp/dto/DriverEarningDto.java
@@ -3,7 +3,7 @@ package com.github.k2ocabhinav.ubercloneapp.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -12,8 +12,8 @@ import java.time.LocalDateTime;
 public class DriverEarningDto {
     private Long id;
     private Long rideId;
-    private Double grossFare;
-    private Double platformCommission;
-    private Double netEarning;
+    private BigDecimal grossFare;
+    private BigDecimal platformCommission;
+    private BigDecimal netEarning;
     private LocalDateTime createdTime;
 }
