+++ ubercloneapp/src/main/java/com/github/k2ocabhinav/ubercloneapp/dto/admin/AdminStatsDto.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -15,8 +16,8 @@ public class AdminStatsDto {
     private Long totalRides;
     private Long completedRides;
     private Long cancelledRides;
-    private Double totalRevenue;
-    private Double platformCommission;
+    private BigDecimal totalRevenue;
+    private BigDecimal platformCommission;
     private Long totalUsers;
     private Long newUsers;
     private Long activeDrivers;
