+++ ubercloneapp/src/main/java/com/github/k2ocabhinav/ubercloneapp/services/PayoutService.java
@@ -4,8 +4,10 @@ import com.github.k2ocabhinav.ubercloneapp.dto.PayoutRequestDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 
+import java.math.BigDecimal;
+
 public interface PayoutService {
-    PayoutRequestDto requestPayout(Double amount);
+    PayoutRequestDto requestPayout(BigDecimal amount);
     PayoutRequestDto approvePayout(Long payoutId, String adminNote);
     PayoutRequestDto rejectPayout(Long payoutId, String adminNote);
     PayoutRequestDto processPayout(Long payoutId);
