+++ ubercloneapp/src/main/java/com/github/k2ocabhinav/ubercloneapp/services/WalletService.java
@@ -5,12 +5,14 @@ import com.github.k2ocabhinav.ubercloneapp.entities.User;
 import com.github.k2ocabhinav.ubercloneapp.entities.Wallet;
 import com.github.k2ocabhinav.ubercloneapp.entities.enums.TransactionMethod;
 
+import java.math.BigDecimal;
+
 public interface WalletService {
-    Wallet addMoneyToWallet(User user, Double amount,
+    Wallet addMoneyToWallet(User user, BigDecimal amount,
                             String transactionId, Ride ride,
                             TransactionMethod transactionMethod);
 
-    Wallet deductMoneyFromWallet(User user, Double amount,
+    Wallet deductMoneyFromWallet(User user, BigDecimal amount,
                                  String transactionId, Ride ride,
                                  TransactionMethod transactionMethod);
