+++ ubercloneapp/src/main/java/com/github/k2ocabhinav/ubercloneapp/strategies/RideFareCalculationStrategy.java
@@ -1,7 +1,8 @@
 package com.github.k2ocabhinav.ubercloneapp.strategies;
 
 import com.github.k2ocabhinav.ubercloneapp.entities.RideRequest;
+import java.math.BigDecimal;
 
 public interface RideFareCalculationStrategy {
-    double calculateFare(RideRequest rideRequest);
+    BigDecimal calculateFare(RideRequest rideRequest);
 }
