+++ ubercloneapp/src/test/java/com/github/k2ocabhinav/ubercloneapp/services/impl/ScheduledRideServiceImplTest.java
@@ -92,7 +92,7 @@ class ScheduledRideServiceImplTest {
             when(riderService.getCurrentRider()).thenReturn(testRider);
             when(modelMapper.map(requestDto, RideRequest.class)).thenReturn(testRideRequest);
             when(rideStrategyManager.rideFareCalculationStrategy()).thenReturn(fareCalculationStrategy);
-            when(fareCalculationStrategy.calculateFare(any())).thenReturn(150.0);
+            when(fareCalculationStrategy.calculateFare(any())).thenReturn(BigDecimal.valueOf(150.0));
             when(rideRequestRepository.save(any(RideRequest.class))).thenReturn(testRideRequest);
             when(modelMapper.map(testRideRequest, ScheduledRideDto.class)).thenReturn(new ScheduledRideDto());
 
@@ -192,7 +192,7 @@ class ScheduledRideServiceImplTest {
             when(rideRequestRepository.findById(1L)).thenReturn(Optional.of(testRideRequest));
             when(riderService.getCurrentRider()).thenReturn(testRider);
             when(rideStrategyManager.rideFareCalculationStrategy()).thenReturn(fareCalculationStrategy);
-            when(fareCalculationStrategy.calculateFare(any())).thenReturn(180.0);
+            when(fareCalculationStrategy.calculateFare(any())).thenReturn(BigDecimal.valueOf(180.0));
             when(rideRequestRepository.save(any(RideRequest.class))).thenReturn(testRideRequest);
             when(modelMapper.map(any(RideRequest.class), eq(ScheduledRideDto.class)))
                     .thenReturn(new ScheduledRideDto());
